# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Book Halal API
  version: 1.0.0
  description: API documentation for the Book Halal reading platform.
servers:
  - url: http://localhost:8090
    description: Local Development Server
security:
  - bearerAuth: []
paths:
...
  /api/books/{id}/ask:
    post:
      summary: Ask a Question About a Book
      description: |
        Use RAG (Retrieval-Augmented Generation) to ask a natural language question about the book's content.
        The service searches relevant passages in the book and answers using DeepSeek AI.
      tags:
        - RAG
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                  example: "Что говорит книга о намазе?"
      responses:
        '200':
          description: AI answer with source references.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '503':
          description: RAG service is unavailable.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AskSource:
      type: object
      properties:
        page:
          type: integer
        text:
          type: string
    AskResponse:
      type: object
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/AskSource'